fix: name each history row for what it is and show agents the attendees
- Every row in the conversations list was labelled "Interaction" with its channel repeated underneath, so a task read "Interaction / task" and a document "Interaction / document"; each row is now named after what it actually is.
- The agent-facing calendar tools described attendees in their own text while returning none, telling an agent that data existed which never arrived.
- Looking attendees up now lives in one place the web app and the agent tools both call, so the two can't drift into answering the same question differently.
